requirements-docker.txt
# 웹 프레임워크 - 단순화된 버전
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# 데이터베이스
sqlalchemy>=2.0.0
psycopg2-binary

# 인증 및 보안
passlib[bcrypt]
python-jose[cryptography]
python-multipart

# 데이터 유효성 검사
pydantic>=2.0.0
email-validator

# 머신러닝 및 컴퓨터 비전 - 최신 호환 가능 버전 사용
torch
torchvision
opencv-python-headless
mediapipe
numpy
pandas
scikit-learn

# AI/ML API
google-generativeai

# 환경 및 설정
python-dotenv

# 유틸리티
Pillow
requests